% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_to_named_list.R
\name{assign_to_named_list}
\alias{assign_to_named_list}
\title{Create a named list of parameters}
\usage{
assign_to_named_list(x, names)
}
\arguments{
\item{x}{Either a vector, a list, or a named list}

\item{names}{names to assign \code{x} to}
}
\value{
A named list object.
}
\description{
Validates parameters to pass either globally or column-wise to a specified
named list
}
\examples{
assign_to_named_list(x = c("item"), names = c("A", "B"))
assign_to_named_list(x = list(c("item1", "item2")), names = c("A", "B"))
assign_to_named_list(x = list(c("item1", "item2")), names = c("A", "B"))
## error when `names(x)` and `names` do not match
assign_to_named_list(
    x = list(A = "item1", B = "item2"),
    names = c("A", "B", "C"))
assign_to_named_list(
    x = list(A = c(x = "item1", y = "item2"), B = "itemB", C = "itemC"),
    names = c("A", "B"))
assign_to_named_list(
    x = list(Q = c(x = "item1", y = "item2"), R = "itemB"),
    names = c("A", "B"))
assign_to_named_list(
    x = list(Q = c(x = "item1", y = "item2"), R = "itemB", S = "itemC"),
    names = c("Q", "R", "S"))

}
\keyword{internal}
