% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_mNIRS.data}
\alias{create_mNIRS.data}
\title{Create an mNIRS Dataframe with Metadata}
\usage{
create_mNIRS.data(data, metadata)
}
\arguments{
\item{data}{A dataframe.}

\item{metadata}{Metadata passed along with the dataframe.
\itemize{
\item nirs_device
\item nirs_columns
\item sample_column
\item event_column
\item sample_rate
\item event_sample_list
\item fit_window
\item display_window
\item end_kinetics_window
}}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Manually add class \code{"mNIRS.data"} and metadata to an existing dataframe.
}
\examples{
## currently implemented metadata
metadata <- list(nirs_device = NULL,
                 nirs_columns = NULL,
                 sample_column = NULL,
                 event_column = NULL,
                 sample_rate = NULL,
                 event_sample_list = NULL,
                 fit_window = NULL,
                 display_window = NULL,
                 end_kinetics_window = NULL)

df <- data.frame(A = 1:3,
                 B = seq(10, 30, 10),
                 C = seq(11, 33, 11))
attributes(df)

nirs_data <- create_mNIRS.data(
    df,
    metadata = list(nirs_columns = c("B", "C"),
                    sample_column = "A",
                    sample_rate = 1)
)
attributes(nirs_data)

}
