% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_mnirs_data}
\alias{create_mnirs_data}
\title{Create a dataframe with metadata}
\usage{
create_mnirs_data(.data, metadata)
}
\arguments{
\item{.data}{A dataframe.}

\item{metadata}{Metadata passed along with the dataframe.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Learned from https://github.com/fmmattioni/whippr/blob/master/R/tbl.R
}
\keyword{internal}
