% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_kinetics.R
\name{find_first_extreme}
\alias{find_first_extreme}
\title{Find the First Extreme Peak or Trough Value in a Numeric Vector}
\usage{
find_first_extreme(y, x = NULL, end_kinetics_span = 20)
}
\arguments{
\item{y}{A numeric vector of the response variable, or the name of the
variable within a dataframe.}

\item{x}{An \emph{optional} numeric vector of the predictor variable, or the name
of the variable within a dataframe. \emph{Defaults} to using the index of
\code{x = seq_along(y)}.}

\item{end_kinetics_span}{(\emph{Default} \code{end_kinetics_span = 20}) A numeric scalar
indicating the local window in units of the predictor variable \code{x} after the
kinetics extreme (peak or trough) value to look for subsequent greater extremes.
The kinetics model will be fit to the data up to the first local extreme with
no subsequent greater extremes within a trailing window in units of \code{x}
defined by \code{end_kinetics_span}, or the limits of the data.}
}
\value{
A list with components:
\item{\code{x}}{The filtered \code{x} values from the origin to the end of the
trailing window defined by \code{end_kinetics_span} past the first extreme value.}
\item{\code{y}}{The filtered \code{y} values from the origin to the end of the
trailing window defined by \code{end_kinetics_span} past the first extreme value.}
\item{\code{extreme}}{A \link[tibble:tibble-package]{tibble} with the \code{x} and \code{y}
coordinates of the first extreme value, as columns.}
}
\description{
Determines the direction of the response of \code{y} (positive or negative) and finds
the first extreme value (peak for positive, or trough for negative) in which no
greater extreme is observed within a trailing window in units of \code{x} defined by
\code{end_kinetics_span}, or the limits of the data.
}
\seealso{
\code{\link[=process_kinetics]{process_kinetics()}}
}
\keyword{internal}
