% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hampel.R
\name{hampel}
\alias{hampel}
\title{Hampel Filter}
\usage{
hampel(x, k, t0 = 3)
}
\arguments{
\item{x}{Numeric vector representing a time series.}

\item{k}{Window length \code{2*k+1} in indices.}

\item{t0}{Threshold, default is 3 (Pearson's rule), see below.}
}
\value{
A list \code{L} with \code{L$y} the corrected time series and
\code{L$ind} the indices of outliers in the \emph{"median absolute deviation"} sense.
}
\description{
Median absolute deviation (MAD) outlier in time series which skips \code{NA}
(modified from \code{\link[pracma:hampel]{pracma::hampel()}})
}
\details{
The \emph{"median absolute deviation"} computation is done in the \verb{[-k...k]}
vicinity of each point at least \code{k} steps away from the end points of the
interval. At the lower and upper end the time series values are preserved.

A high threshold makes the filter more forgiving, a low one will declare
more points to be outliers. \code{t0 <- 3} (the default) corresponds to Ron
Pearson's 3 sigma edit rule, \code{t0 <- 0} to John Tukey's median filter.

This modified Hampel filter records the index of \code{NA} values, omits \code{NA},
applies the filter, then re-inserts those \code{NA} back to their original
indices, preserving the original input data indices and replacing outliers
as intended.
}
\examples{
set.seed(8421)
x <- numeric(1024)
z <- rnorm(1024)
x[1] <- z[1]
for (i in 2:1024) {
    x[i] <- 0.4*x[i-1] + 0.8*x[i-1]*z[i-1] + z[i]
}
x[c(3, 8)] <- NA
omad <- hampel(x, k=20)

## Not run:
plot(1:1024, x, type="l")
points(omad$ind, x[omad$ind], pch=21, col="darkred")
grid()
## End(Not run)

}
\seealso{
\code{\link[pracma:hampel]{pracma::hampel()}}
}
