% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/handle_missing_data.R
\name{handle_missing_data}
\alias{handle_missing_data}
\title{Handle missing data in mNIRS Data}
\usage{
handle_missing_data(
  .data,
  handle_missing_values = c("interpolate", "none"),
  ...
)
}
\arguments{
\item{.data}{A dataframe with mNIRS metadata.}

\item{handle_missing_values}{Indicates how to handle missing data.
\itemize{
\item \emph{"interpolate"} (the default) will interpolate between existing
values.
\item \emph{"none"} will pass through missing data. This may introduce errors
further along in processing.
}}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Linearly interpolate across missing mNIRS data.
}
\details{
Additional details...
}
