% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mNIRS-package.R
\docType{package}
\name{mNIRS-package}
\alias{mNIRS}
\alias{mNIRS-package}
\title{mNIRS: Methods for Muscle Near-Infrared Spectroscopy Processing and Analysis}
\description{
Methods for importing, processing, and analysis data from muscle near-infrared spectroscopy (mNIRS) devices.
}
\author{
\strong{Maintainer}: Jem Arnold \email{jem.arnold@ubc.com} (\href{https://orcid.org/0000-0003-3908-9447}{ORCID})

}
\keyword{internal}
