% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.mNIRS.kinetics.R
\name{mNIRS.kinetics-methods}
\alias{mNIRS.kinetics-methods}
\alias{print.mNIRS.kinetics}
\title{Methods for mNIRS.kinetics objects}
\usage{
\method{print}{mNIRS.kinetics}(x, ...)
}
\arguments{
\item{x}{object of class \code{mNIRS.kinetics} as returned from
\code{\link[=process_kinetics]{process_kinetics()}}}

\item{...}{further arguments passed through, see description of return value
for details.}
}
\value{
\describe{
\item{\code{print}}{Prints the mNIRS kinetics model summary}
\item{\code{plot}}{Returns a plot of mNIRS kinetics}
}
}
\description{
Methods defined for objects returned from \code{\link[=process_kinetics]{process_kinetics()}}.
}
