% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoexponential.R
\name{monoexp_init}
\alias{monoexp_init}
\title{Initiate Self-Starting \code{nls} Monoexponential Model}
\usage{
monoexp_init(mCall, data, LHS, ...)
}
\arguments{
\item{mCall}{A matched call to the function \code{model}.}

\item{data}{A data frame in which to interpret the variables in \code{mCall}.}

\item{LHS}{The expression from the left-hand side of the model formula in the
call to \code{nls}.}

\item{...}{Additional arguments.}
}
\value{
Initial starting estimates for parameters in the model called by
\code{\link[=SSmonoexp]{SSmonoexp()}}.
}
\description{
This function should return initial values for the parameters in model.
}
\seealso{
\link[stats:nls]{nls}, \link[stats:selfStart]{selfStart},
\link[=SSmonoexp]{SSmonoexp}, \link[=monoexponential]{monoexponential}
}
