% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoexponential.R
\name{monoexponential}
\alias{monoexponential}
\title{Monoexponential Function with 4 Parameters}
\usage{
monoexponential(x, A, B, TD, tau)
}
\arguments{
\item{x}{A numeric predictor variable at which to evaluate the response
variable \code{y}.}

\item{A}{A numeric parameter for the starting (baseline) value of the response
variable \code{y}.}

\item{B}{A numeric parameter for the ending (asymptote) value of the response
variable \code{y}.}

\item{TD}{A numeric parameter for the time delay before exponential inflection
of the curve, in units of the predictor variable \code{x}.}

\item{tau}{A numeric parameter for the time constant \verb{tau (𝜏)} of the
exponential curve, in units of the predictor variable \code{x}. \code{tau} is equal to
the reciprocal of \code{k} (\code{tau = 1/k}), where \code{k} is the rate constant of the
same function.}
}
\value{
A numeric vector for the response variabel \code{y} of the same length as
the predictor variable \code{x}.
}
\description{
Calculate a four-parameter monoexponential curve.
}
\examples{
set.seed(13)
x <- seq(0, 60, by = 2)
A <- 10; B <- 100; TD <- 15; tau <- 8
y <- monoexponential(x, A, B, TD, tau) + rnorm(length(x), 0, 3)
data <- data.frame(x, y)

model <- nls(y ~ SSmonoexp(x, A, B, TD, tau), data = data)
model

\dontrun{
plot(x, y)
lines(x, y)
points(x, y)
lines(x, fitted(model), col = "red")
}

}
