% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_dataframe.R
\name{normalise_dataframe}
\alias{normalise_dataframe}
\title{Normalise data range}
\usage{
normalise_dataframe(data, nirs_columns = list(), normalise_range = c(0, 100))
}
\arguments{
\item{data}{A dataframe containing mNIRS data.}

\item{nirs_columns}{A \code{list()} of character vectors indicating the column
names for data signals to be shifted. Columns grouped together in a vector
will preserve relative scaling. Separate columns will shift to their own
specified values. Should match column names in the dataframe exactly. See
\emph{Details}.}

\item{normalise_range}{A numeric vector in the form \code{c(min, max)},
indicating the range of values to which the specified mNIRS data
columns will be re-scaled.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Normalise the range of mNIRS signals while preserving the relative
scaling across columns.
}
\details{
\code{nirs_columns} can be used to group data columns together to preserve
relative scaling across mNIRS signals.
\itemize{
\item \code{list(A, B, C)} will rescale each column separately. Relative values will
be shifted between columns.
\item \code{list(c(A, B, C))} will rescale all columns together to a common range.
value. Relative scaling is preserved across the group of columns.
\item \code{list(c(A, B), c(C, D))} will rescale columns \code{A} and \code{B} to a common
range, and columns \code{C} and \code{D} to a separate common range. This is a way to
create multiple groups of data columns where relative scaling is preserved
within groups, but not across groups of mNIRS signals.
}
}
