% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_slope.R
\name{peak_slope}
\alias{peak_slope}
\title{Calculate Peak Positive or Negative Slope}
\usage{
peak_slope(
  y,
  x = seq_along(y),
  width,
  align = c("center", "left", "right"),
  na.rm = FALSE
)
}
\arguments{
\item{y}{A numeric vector of the response variable.}

\item{x}{A numeric vector of the predictor variable, defaults to using the
index of \code{x = seq_along(y)}.}

\item{width}{A numeric scalar defining the window width in units of \code{x}
for rolling calculations.}

\item{align}{Specifies the window alignment of \code{width} as \emph{"center"}
(the \emph{default}), \emph{"left"}, or \emph{"right"}. Where \emph{"left"} is \emph{forward looking},
and \emph{"right"} is \emph{backward looking} by the window \code{width} from the current
sample.}

\item{na.rm}{A logical indicating how missing data will be handled. \code{FALSE}
(the \emph{default}) will perform rolling slopes with complete case analysis and
return the peak of rolling slopes where all local \code{y} samples are valid.
\code{TRUE} will return the peak of rolling slopes where the local target \code{y}
sample is valid.}
}
\value{
A list of class \code{mNIRS.kinetics} with components \code{L$...}:
\item{\code{x}}{The \code{x} position of the local peak slope.}
\item{\code{y}}{The \code{y} value predicted from linear regression at the local
peak slope.}
\item{\code{slope}}{The slope value in units of \code{y/x}.}
\item{\code{x_fitted}}{The range of \code{x} values given by \code{width} corresponding
to the local peak slope.}
\item{\code{y_fitted}}{The range of \code{y} values given by \code{width} predicted
from linear regression, corresponding to the local peak slope.}
}
\description{
Computes the peak directional linear regression slope for upward or downward
trending response variable \code{y}, respectively using \code{\link[=rolling_slope]{rolling_slope()}} within
a moving window.
}
\details{
Uses the least squares formula on complete case analysis to calculate
local slopes within a rolling window along \code{x} specified by \code{width} in units
of \code{x}. i.e. \code{x = 10} would include samples within a 10-second window for
time series data.

The direction of \code{y} (upward or downward) will determine whether a positive
or negative peak slope is returned. Peak positive slope will be returned for
an overall slope >= 0.
}
\examples{
y <- c(1, 3, 2, 5, 8, 7, 9, 12, 11, 14)
peak_slope(y, width = 3)

y <- c(14, 11, 12, 9, 7, 8, 5, 2, 3, 1)
peak_slope(y, width = 3)

}
