% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mNIRS.data.R
\name{plot.mNIRS.data}
\alias{plot.mNIRS.data}
\title{Plot mNIRS.data objects}
\usage{
\method{plot}{mNIRS.data}(data, na.omit = FALSE, ...)
}
\arguments{
\item{data}{Object of class \code{mNIRS.data} as returned from \code{\link[=create_mNIRS.data]{create_mNIRS.data()}}}

\item{na.omit}{A logical indicating whether missing data (\code{NA}) should be omitted
(\code{TRUE}) for better display of disconnected lines, or included (\code{FALSE}, the
default) to better recognise where data are missing.}

\item{...}{Additional arguments (\emph{currently not used}).}
}
\value{
A \link[ggplot2:ggplot]{ggplot2} object.
}
\description{
Create a simple plot for objects returned from \code{\link[=create_mNIRS.data]{create_mNIRS.data()}}.
}
