% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data.R
\name{prepare_kinetics_data}
\alias{prepare_kinetics_data}
\title{Prepare mNIRS Data for Kinetics Analysis}
\usage{
prepare_kinetics_data(
  data,
  event_index = NULL,
  event_sample = NULL,
  event_label = NULL,
  fit_window = c(30, 180),
  display_window = NULL,
  group_events = list("distinct", "ensemble"),
  ...
)
}
\arguments{
\item{data}{A dataframe of class \code{"mNIRS.data"}.}

\item{event_index}{(\emph{Optional}). A numeric vector indicating the starting row
indices of kinetics events. i.e., to identify the start of kinetic events by
row number.}

\item{event_sample}{(\emph{Optional}). A numeric vector corresponding to values of
\code{sample_column} indicating the start of kinetic events. i.e., to identify
the start of kinetic events by time value or sample number.}

\item{event_label}{(\emph{Optional}). A character vector corresponding to values of
\code{event_column} indicating the start of kinetics events. i.e., to identify
the start of kinetic events by an event string such as \emph{"end work"}.}

\item{fit_window}{A two-element numeric vector in the form \code{c(before, after)}
in units of the \code{sample_column}, defining the window before and after the
kinetics events to include in the model fitting process
(\emph{default = c(30, 180)}).}

\item{display_window}{(\emph{Optional}). A two-element numeric vector in the form
\code{c(before, after)} in units of the \code{sample_column}, defining the window before
and after the kinetics events to include for display, but not for model
fitting.}

\item{group_events}{Indicates how kinetics events should be analysed. Typically
either \emph{"distinct"} (\emph{the default}) or \emph{"ensemble"}, but can be manually
specified (see \emph{Details}).}

\item{...}{Additional arguments.}
}
\value{
A list of \link[tibble:tibble-package]{tibbles} of class \code{mNIRS.data}
with metadata available with \code{attributes()}.
}
\description{
Processes a dataframe of class \code{"mNIRS.data"} for kinetics analysis.
}
\details{
\code{group_events} indicates how kinetics events should be analysed, either
separately, or grouped and ensemble averaged similar to oxygen uptake kinetics.

\describe{
\item{\code{group_events = "distinct"}}{Will prepare a list of unique dataframes
for each kinetics event (\emph{default}).}
\item{\code{group_events = "ensemble"}}{Will prepare one dataframe with the
ensemble-averaged data from all mNIRS kinetics events.}
\item{\code{group_events = list(c(1, 2), c(3, 4))}}{Will group kinetic events
together in sequence of appearance, and prepare a list of ensemble-averaged
dataframes for each group. Any kinetic events detected in the data but
not explicitly defined here will return as a distinct dataframe.}
}
}
