% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data_old.R
\name{prepare_kinetics_data_old}
\alias{prepare_kinetics_data_old}
\title{Prepare Dataframes for mNIRS Kinetics Analysis}
\usage{
prepare_kinetics_data_old(
  .data,
  fit_baseline_window = 30,
  fit_kinetics_window = 180,
  display_baseline_window = 30,
  display_kinetics_window = 180,
  end_kinetics_window = NULL,
  multiple_kinetics_events = c("distinct", "ensemble"),
  ...
)
}
\arguments{
\item{.data}{The data retrived \code{\link[=process_data]{process_data()}}.}

\item{fit_baseline_window}{A numeric scalar specifying the number of
samples preceding to the kinetics events to include as baseline.}

\item{fit_kinetics_window}{A numeric scalar specifying the number of
samples following to the kinetics events to include for kinetics modelling.}

\item{display_baseline_window}{A numeric scalar specifying the number of
samples preceding to the kinetics events to include for display (will not
necessarily be modelled).}

\item{display_kinetics_window}{A numeric scalar specifying the number of
samples following to the kinetics events to include for display (will not
necessarily be modelled).}

\item{end_kinetics_window}{A numeric scalar specifying the number of
samples in which to look for a peak or nadir value indicating the kinetics
plateau.}

\item{multiple_kinetics_events}{Indicates whether kinetics events should
be analysed separately or together.
\itemize{
\item \emph{"distinct"} (the default) will prepare a list of unique dataframes for
each kinetics event.
\item \emph{"ensemble"} will prepare one dataframe with the ensemble-averaged mNIRS
data from all kinetics events (similar to pulmonary VO2 kinetics).
\item Alternatively, a list of numeric vectors can be specified to
ensemble-average groups of kinetics events together. Such as
\code{multiple_kinetics_events = list(c(1, 2), c(3, 4))} for two sets of
repeated bouts.
}}

\item{...}{Additional arguments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
This function retrieves processed data from \code{\link[=process_data]{process_data()}} with manually
identified kinetics events. Kinetics parameters are defined to extract
data excerpts to fit those kinetics events. Each kinetics event can be
analysed separately, or ensemble-averaged and analysed together.
}
\details{
TODO
}
