% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_kinetics.R
\name{process_kinetics}
\alias{process_kinetics}
\title{Process Kinetics}
\usage{
process_kinetics(
  x,
  idx = "index",
  method = c("monoexponential", "logistic", "half-time", "peak-slope"),
  end_kinetics_width,
  ...
)
}
\arguments{
\item{x}{A numeric vector of mNIRS data.}

\item{end_kinetics_width}{A numeric scalar specifying the number of
samples in which to look for a peak or nadir value indicating the kinetics
plateau.}

\item{...}{Additional arguments (\emph{currently not used}).}
}
\value{
A list \code{L} with \code{L$y} the kinetics time series and \code{L$x} the
raw data...
}
\description{
Process Kinetics
}
\details{
...
}
