% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read data from mNIRS device}
\usage{
read_data(
  file_path,
  nirs_columns,
  sample_column = NULL,
  event_column = NULL,
  .keep_all = FALSE,
  ...
)
}
\arguments{
\item{file_path}{The local file path including extension (either \emph{".xlsx"},
\emph{".xls"}, or \emph{".csv"}) to import.}

\item{nirs_columns}{A character vector indicating the mNIRS data columns
to import from the target file. Must match exactly. Optionally, a named
vector to rename the columns.}

\item{sample_column}{(optional) A character scalar indicating the name of
the time or sample data column. Must match exactly. Default is \code{NULL}.
Optionally, a named scalar to rename the column.}

\item{event_column}{(optional) A character scalar indicating the name of
the event or lap data column. Must match exactly. Default is \code{NULL}.
Optionally, a named scalar to rename the column.}

\item{.keep_all}{A logical, if \code{FALSE} (the default) will only include
the target data columns. If \code{TRUE} will include all columns detected from
the file.}

\item{...}{Additional arguments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
This function reads and imports mNIRS data from file, and returns a
dataframe.
}
\details{
Additional arguments will accept \code{sample_rate} as an integer scalar for the
data sample rate in Hz. This is required for certain functions. \code{sample_rate}
will be estimated based on the mean difference between values in the
\code{sample_column}. If this column contains decimal or date-time values,
it should correctly estimate the sample rate in Hz. If this column
contains integer sample numbers, it will incorrectly estimate \code{sample_rate}
as 1 Hz, and the correct value should be entered manually.

A method has been included to correctly identify \emph{Artinis Oxysoft} sample
rates, from files exported in English language.
}
