% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_invalid.R
\name{replace_invalid}
\alias{replace_invalid}
\title{Replace Invalid Values}
\usage{
replace_invalid(x, values, width, return = c("NA", "median"))
}
\arguments{
\item{x}{A numeric vector}

\item{values}{A numeric vector of invalid values to be replaced, e.g.
\code{values = c(0, 100, 102.3)}.}

\item{width}{A numeric scalar for the window length of \verb{(2 · width + 1)} samples.}

\item{return}{Indicates whether invalid values should be replaced with \code{NA}
(the \emph{default}) or the local \emph{"median"} value.}
}
\value{
A numeric vector of filtered data.
}
\description{
Detect specific values such as \code{c(0, 100, 102.3)} in vector data and replaces
with \code{NA} or the local median value.
}
\details{
Useful to overwrite known invalid/nonsense values, such as \code{0},
\code{100}, or \code{102.3}.

\emph{TODO: allow for overwriting all values greater or less than known values.}
}
\examples{
set.seed(13)
(x <- sample.int(10, 20, replace = TRUE))
(y <- replace_invalid(x, values = c(1, 10), width = 5))

}
\seealso{
\code{\link[pracma:hampel]{pracma::hampel()}}
}
