% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing.R
\name{replace_missing}
\alias{replace_missing}
\title{Replace Missing Values}
\usage{
replace_missing(x, method = c("linear", "locf"), na.rm = TRUE, maxgap = Inf)
}
\arguments{
\item{x}{A numeric vector.}

\item{method}{Indicates how to replace missing data.
\describe{
\item{\code{"linear"}}{Replaces each \code{NA} with linear interpolation via
\code{\link[zoo:na.approx]{zoo::na.approx()}} (the \emph{default}).}
\item{\code{"locf"}}{Stands for 'last observation carried forward'. Replaces each
\code{NA} with the most recent non-\code{NA} value prior to it, or following it for
leading \code{NA}s.}
\item{\code{"omit"}}{\verb{<under development>}}
}}

\item{na.rm}{A logical indicating whether leading and/or trailing \code{NA}s should
be extrapolated/filled (\code{TRUE}, the \emph{default}) or ignored (\code{FALSE}).}

\item{maxgap}{A numeric scalar for the maximum number of consecutive \code{NA}s
to fill. Any longer gaps will be left unchanged. \emph{Default} is \code{Inf}.}
}
\value{
A numeric vector of filtered data.
}
\description{
Detect missing values in vector data and replace via methods from the
\link[zoo:na.approx]{zoo} package.
}
\details{
This function will also replace \code{c(NaN, Inf, -Inf)}.

\describe{
\item{\code{method = "linear"}}{\code{na.rm = TRUE} will extrapolate over leading and
trailing \code{NA}s by applying \code{rule = 2} (see \code{\link[stats:approxfun]{stats::approx()}}).
\code{na.rm = FALSE} will return leading/trailing \code{NA}s by applying \code{rule = 1}.}
\item{\code{method = "locf"}}{\code{na.rm = TRUE} will fill NOCB 'next observation
carried backward over leading \code{NA}s. \code{na.rm = FALSE} will return leading
\code{NA}s.}
}
}
\examples{
(x <- c(1, 2, NA, Inf, 5, 6, NA))
replace_missing(x, method = "linear", na.rm = FALSE)
replace_missing(x, method = "linear", na.rm = TRUE)
replace_missing(x, method = "locf")

}
\seealso{
\code{\link[zoo:na.locf]{zoo::na.locf()}}, \code{\link[zoo:na.approx]{zoo::na.approx()}}
}
