% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_missing_values.R
\name{replace_missing_data}
\alias{replace_missing_data}
\title{Replace Missing Values}
\usage{
replace_missing_data(
  x,
  method = c("locf", "linear", "spline"),
  rule = 2,
  na.rm = FALSE,
  maxgap = Inf,
  ...
)
}
\arguments{
\item{x}{A numeric vector of mNIRS data.}

\item{method}{Indicates how to replace missing data.
\itemize{
\item \emph{"locf"} Stands for 'last observation carried forward'. Replaces each \code{NA}
with the most recent non-\code{NA} value prior to it.
\item \emph{"linear"} Replaces each \code{NA} via linear interpolation.
\item \emph{"spline"} Replaces each \code{NA} with cubic spline interpolation.
}}

\item{rule}{An integer of length 1 or 2 describing how interpolation is to
take place outside the interval \code{\link{min(x), max(x)}}. If \code{rule = 1} then \code{NA}s
are returned for such points and if \code{rule = 2}, the value at the closest
data extreme is used. Use, e.g., \code{rule = 2:1}, if the left and right side
extrapolation should differ.}

\item{na.rm}{A logical. If the result of the interpolation still results in
leading and/or trailing \code{NA}s, should these be removed (using na.trim)?}

\item{maxgap}{A numeric scalar for the maximum number of consecutive \code{NA}s
to fill. Any longer gaps will be left unchanged.}

\item{...}{Additional arguments.}
}
\value{
A list \code{L} with \code{L$y} the corrected time series and
\code{L$idx} the indices of the values replaced.
}
\description{
Detect missing values within mNIRS vector data and replace via methods from
the \code{{zoo}} package.
}
\details{
For \emph{"locf"}, if there are no earlier non-\code{NA}s, then the \code{NA} is either
omitted (if \code{na.rm = TRUE}) or it is not replaced (if \code{na.rm = FALSE}).

...
}
