% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_outliers.R
\name{replace_outliers}
\alias{replace_outliers}
\title{Replace Fixed Values and Outliers from mNIRS Data}
\usage{
replace_outliers(.x, fixed_values = NULL, outliers = FALSE, k = 20, ...)
}
\arguments{
\item{.x}{A numeric vector of mNIRS data.}

\item{fixed_values}{(optional) Either \code{NULL} (the default) or a
numeric vector of values to be overwritten. e.g. \code{fixed_values = c(0, 100)}.}

\item{outliers}{A logical indicating whether local outliers should
be replaced using a Hampel filter \code{\link[=hampel]{hampel()}}.}

\item{k}{A numeric scalar for the window length of \verb{(2 · k + 1)} indices,
for local outlier detection and the median value replacement.}

\item{...}{Additional arguments.}
}
\value{
A numeric vector.
}
\description{
Detect specified values such as \code{c(0, 100)}, and local outliers with
\code{\link[=hampel]{hampel()}} within mNIRS vector data, and replaces with the local median value.
}
\details{
Additional arguments will accept \verb{t0, na.rm, return} for \code{\link[=hampel]{hampel()}}.
}
