% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_data.R
\name{resample_data}
\alias{resample_data}
\title{Re-sample Dataframe}
\usage{
resample_data(
  data,
  sample_channel = NULL,
  sample_rate = NULL,
  resample_rate = NULL,
  resample_time = NULL,
  verbose = TRUE
)
}
\arguments{
\item{data}{A dataframe.}

\item{sample_channel}{A character scalar indicating the time or sample data
column. Must match \code{data} column names exactly. Will be taken from metadata
if not defined explicitly.}

\item{sample_rate}{A numeric scalar for the sample rate in Hz. Will be taken
from metadata if not defined explicitly.}

\item{resample_rate}{An \emph{optional} numeric scalar indicating the desired
output sample rate (in Hz) to convert the dataframe.}

\item{resample_time}{An \emph{optional} numeric scalar indicating the desired
sample time (in seconds) to convert the dataframe.}

\item{verbose}{A logical. \code{TRUE} (the \emph{default}) will return warnings and
messages which can be used for data error checking. \code{FALSE} will silence these
messages. Errors will always be returned.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mnirs.data} with
metadata available with \code{attributes()}.
}
\description{
Increase or decrease the number of observations in a dataframe using linear
interpolation for up- and down-sampling.
}
\details{
\code{sample_channel} and \code{sample_rate} will be taken from metadata for a dataframe
of class \verb{"mnirs.data} which has been processed with \code{{mnirs}}, if not specified
explicitly.

Otherwise, \code{sample_rate} will be estimated from the values in \code{sample_channel}.
However, this may return unexpected values, and it is safer to define
\code{sample_rate} explicitly.
}
