% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dataframe.R
\name{resample_dataframe}
\alias{resample_dataframe}
\title{Resample Data}
\usage{
resample_dataframe(
  data,
  sample_rate,
  resample_rate = NULL,
  resample_time = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe containing mNIRS data.}

\item{...}{Additional arguments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Resamples mNIRS data using either time-weighted or linear
interpolation
}
\details{
...
}
