% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mnirs.data.R
\name{scale_colour_mnirs}
\alias{scale_colour_mnirs}
\alias{scale_fill_mnirs}
\title{Scales for custom mNIRS palette}
\usage{
scale_colour_mnirs(...)

scale_fill_mnirs(...)
}
\arguments{
\item{...}{Arguments passed to discrete_scale.}
}
\description{
Scales for custom mNIRS palette
}
\seealso{
\code{\link[=mnirs_palette]{mnirs_palette()}}
}
