% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_data.R
\name{shift_data}
\alias{shift_data}
\title{Shift Data Range}
\usage{
shift_data(
  data,
  nirs_channels = list(),
  sample_channel = NULL,
  shift_to = NULL,
  shift_by = NULL,
  span = 0,
  position = c("minimum", "maximum", "first")
)
}
\arguments{
\item{data}{A dataframe.}

\item{nirs_channels}{A \code{list()} of character vectors indicating the column
names for data channels to be shifted (see \emph{Details}).
\describe{
\item{\code{nirs_channels = list("A", "B", "C")}}{will shift each channel
separately, losing relative scaling.}
\item{\code{nirs_channels = list(c("A", "B", "C"))}}{will shift all channels
together, preserving relative scaling.}
\item{\code{nirs_channels = list(c("A", "B"), c("C", "D"))}}{will shift channels
\code{A} and \code{B} together, and channels \code{C} and \code{D} together, preserving
relative scaling within, but not across each group.}
}}

\item{sample_channel}{A character scalar indicating the time or sample data
channel. Must match \code{data} column names exactly. Will be taken from metadata
if not defined explicitly.}

\item{shift_to}{A positive or negative numeric scalar in units of \code{nirs_channels}
to which the specified data channels will be shifted.}

\item{shift_by}{An \emph{optional} positive or negative numeric scalar by which the
data signals can be shifted by a set amount, in units of the \code{nirs_channels}.}

\item{span}{A numeric scalar in units of \code{sample_channel}, defining the range
over which the \code{position} is determined. \code{span = 0} takes the single
minimum, maximum, or first value. \code{span = 30} would use the mean of the samples
within the lowest, highest, or first \code{30} units of the \code{sample_channel}.}

\item{position}{Indicates how to shift values.
\describe{
\item{\code{"minimum"}}{will shift selected channels' minimum values
to the specified \code{shift_to} value (the \emph{default}).}
\item{\code{"maximum"}}{will shift selected channels by their maximum values.}
\item{\code{"first"}}{will shift selected channels by their first values.}
}}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Move the range of data channels in a dataframe up or down, while preserving
the absolute dynamic range and/or relative scaling across channels. e.g. shift
data range to positive values, or shift the mean of the first X number of
samples in a recording to zero.
}
\details{
\code{nirs_channels = list()} can be used to group data channels to preserve
absolute or relative scaling. Channels grouped together in a vector
will preserve relative scaling across those channels. Should match column names
in the dataframe exactly.

\describe{
\item{\code{nirs_channels = list("A", "B", "C")}}{will shift each channel
separately. Absolute dynamic range for each data channel is preserved, but
relative scaling will be lost between data channels.}
\item{\code{nirs_channels = list(c("A", "B", "C"))}}{will shift all channels
together. Absolute dynamic range and relative scaling are both preserved
across the group of data channels.}
\item{\code{nirs_channels = list(c("A", "B"), c("C", "D"))}}{will shift channels
\code{A} and \code{B} together, and channels \code{C} and \code{D} together. Absolute dynamic
range and relative scaling are preserved within each group, but not across
groups of data channels.}
}

Channels not explicitly specified will not be shifted, but will be passed
through to the output dataframe.

If both \code{shift_to} and \code{shift_by} are defined, only \code{shift_to} will be used.
}
