% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shift_dataframe.R
\name{shift_dataframe}
\alias{shift_dataframe}
\title{Shift Data Range}
\usage{
shift_dataframe(
  data,
  nirs_columns = list(),
  shift_to = 0,
  position = c("minimum", "maximum", "first"),
  mean_samples = 1,
  shift_by = NULL
)
}
\arguments{
\item{data}{A dataframe containing mNIRS data.}

\item{nirs_columns}{A \code{list()} of character vectors indicating the column
names for data signals to be shifted. Columns grouped together in a vector
will preserve relative scaling. Separate columns will shift to their own
specified values. Should match column names in the dataframe exactly. See
\emph{Details}.}

\item{shift_to}{A numeric scalar to which the specified data columns
will be shifted to.}

\item{position}{Indicates how to shift values.
\describe{
\item{\code{"minimum"}}{\emph{(default)} will shift selected columns' minimum values
to the specified \code{shift_to} value.}
\item{\code{"maximum"}}{will shift selected columns by their maximum values.}
\item{\code{"first"}}{will shift selected columns by their first existing
values.}
}}

\item{mean_samples}{An integer scalar representing the number of samples
over which the \code{position} is determined. e.g., \code{mean_samples = 1} looks for
the single minimum, maximum, or first value. \code{mean_samples = 30} would
use the mean of the lowest, highest, or first \code{30} samples.}

\item{shift_by}{\emph{(optional)} specified data columns can be shifted by a set
amount.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Shift the data range of mNIRS signals (e.g. shift to positive values) while
preserving the absolute dynamic range and relative scaling across columns.
}
\details{
\code{nirs_columns} can be used to group data columns together to preserve
relative scaling across mNIRS signals.

\describe{
\item{\code{list(A, B, C)}}{will shift each column separately within their own
values. Absolute dynamic range is preserved, but relative values will be
shifted between columns.}
\item{\code{list(c(A, B, C))}}{will shift all columns together to a common
specified value. Absolute dynamic range and relative scaling are both
preserved across the group of columns.}
\item{\code{list(c(A, B), c(C, D))}}{will shift columns \code{A} and \code{B} to a
common value, and columns \code{C} and \code{D} to a separate common value.
This is a way to create multiple groups of data columns where relative
scaling is preserved within groups, but not across groups of mNIRS
signals.}
}
}
