% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signif_trailing.R
\name{signif_trailing}
\alias{signif_trailing}
\title{Print a numeric value as character with trailing zeroes}
\usage{
signif_trailing(x, digits = 2, format = c("digits", "signif"))
}
\arguments{
\item{x}{A numeric vector.}

\item{digits}{An integer scalar specifying the number of digits to preserve.}

\item{format}{Indicates how to treat \code{digits}. Either the desired number of
digits after the decimal point (\code{format = "digits"}) or significant digits
(\code{format = "signif"}).}
}
\value{
A character vector.
}
\description{
Converts a numeric to a character object to preserve trailing zeroes
for display.
}
\examples{
x <- c(1, 1.3, 1.35)
signif_trailing(x, digits = 2, format = "digits")
signif_trailing(x, digits = 2, format = "signif")

}
\seealso{
\code{\link[=formatC]{formatC()}}, \code{\link[=round]{round()}}, \code{\link[=signif]{signif()}}
}
\keyword{internal}
