% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_slope.R
\name{slope}
\alias{slope}
\title{Calculate the Slope of a Numeric Vector}
\usage{
slope(y, x = seq_along(y))
}
\arguments{
\item{y}{A numeric vector of the response variable.}

\item{x}{A numeric vector of the predictor variable, defaults to using the
index of \code{x = seq_along(y)}.}
}
\value{
A numeric slope value of \code{y/x}.
}
\description{
Computes the linear regression slope for a numeric response variable \code{y}.
}
\details{
Uses the least squares formula on complete case analysis. A single
valid observation will return \code{NA}.
}
\examples{
\dontrun{
y <- c(1, 3, 2, 5, 8, 7, 9, 12, 11, 15, 14, 17, 18)
slope(y)

y <- c(1, 3, NA, 5, 8, 7, 9, 12, NA, NA, NA, 17, NA)
slope(y)
}

}
\keyword{internal}
