% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoexponential.R
\name{monoexp_init}
\alias{monoexp_init}
\alias{SSmonoexp}
\title{Initiate Self-Starting \code{nls} Monoexponential Model}
\usage{
monoexp_init(mCall, data, LHS, ...)

SSmonoexp(x, A, B, TD, tau)
}
\arguments{
\item{mCall}{A matched call to the function \code{model}.}

\item{data}{A data frame in which to interpret the variables in \code{mCall}.}

\item{LHS}{The expression from the left-hand side of the model formula in the
call to \code{nls}.}

\item{...}{Additional arguments.}

\item{x}{A numeric predictor variable at which to evaluate the model.}

\item{A}{A numeric parameter for the starting (baseline) value of the response
variable \code{y}.}

\item{B}{A numeric parameter for the ending (asymptote) value of the response
variable \code{y}.}

\item{TD}{A numeric parameter for the time delay before exponential inflection
of the curve, in units of the predictor variable \code{x}.}

\item{tau}{A numeric parameter for the time constant (𝜏) of the exponential
curve, in units of the predictor variable \code{x}.}
}
\value{
Initial starting estimates for parameters in the model called by
\code{\link[=SSmonoexp]{SSmonoexp()}}.

A numeric vector for the response variabel \code{y} of the same length as
the predictor variable \code{x}. Returned from the expression
\code{ifelse(x <= TD, A, A + (B - A) * (1 - exp((TD - x) / tau)))}.
}
\description{
Return initial values for the parameters in a \code{selfStart} model.

Create initial coefficient estimates for a \code{selfStart} model for the
four-parameter \code{\link[=monoexponential]{monoexponential()}} function. For the parameters \code{A}, \code{B},
\code{TD}, and \code{tau}.
}
\examples{
set.seed(13)
x <- seq(0, 60, by = 2)
A <- 10
B <- 100
TD <- 15
tau <- 8
y <- monoexponential(x, A, B, TD, tau) + rnorm(length(x), 0, 3)
data <- data.frame(x, y)

model <- nls(y ~ SSmonoexp(x, A, B, TD, tau), data = data)
model
fitted <- fitted(model)

plot(x, y)
lines(x, y)
points(x, y)
lines(x, fitted, col = "red")

}
\seealso{
\code{\link[stats:nls]{stats::nls()}}, \code{\link[stats:selfStart]{stats::selfStart()}}, \code{\link[=SSmonoexp]{SSmonoexp()}}, \code{\link[=monoexponential]{monoexponential()}}

\code{\link[stats:nls]{stats::nls()}}, \code{\link[stats:selfStart]{stats::selfStart()}}, \code{\link[=monoexponential]{monoexponential()}}
}
