% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_mnirs_data}
\alias{create_mnirs_data}
\title{Create a Dataframe with Metadata}
\usage{
create_mnirs_data(data, metadata)
}
\arguments{
\item{data}{A dataframe.}

\item{metadata}{Metadata passed along with the dataframe.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Description
}
\keyword{internal}
