% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_mnirs_data}
\alias{create_mnirs_data}
\title{Create an mNIRS Dataframe with Metadata}
\usage{
create_mnirs_data(data, metadata)
}
\arguments{
\item{data}{A dataframe.}

\item{metadata}{Metadata passed along with the dataframe.
\itemize{
\item nirs_device
\item nirs_channels
\item sample_channel
\item event_channel
\item sample_rate
\item event_sample_list
\item fit_span
\item display_span
\item end_kinetics_span
}}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mnirs.data} with
metadata available with \code{attributes()}.
}
\description{
Manually add class \code{"mnirs.data"} and metadata to an existing dataframe.
}
\examples{
## currently implemented metadata
metadata <- list(nirs_device = NULL,
                 nirs_channels = NULL,
                 sample_channel = NULL,
                 event_channel = NULL,
                 sample_rate = NULL,
                 event_sample_list = NULL,
                 fit_span = NULL,
                 display_span = NULL,
                 end_kinetics_span = NULL)

df <- data.frame(A = 1:3,
                 B = seq(10, 30, 10),
                 C = seq(11, 33, 11))
attributes(df)

nirs_data <- create_mnirs_data(
    df,
    metadata = list(nirs_channels = c("B", "C"),
                    sample_channel = "A",
                    sample_rate = 1)
)
attributes(nirs_data)

}
\keyword{internal}
