% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{create_mnirs_data}
\alias{create_mnirs_data}
\title{Create an mNIRS Dataframe with Metadata}
\usage{
create_mnirs_data(data, metadata)
}
\arguments{
\item{data}{A dataframe.}

\item{metadata}{Metadata passed along with the dataframe.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Used to manually add class "mNIRS.data" and metadata to an existing dataframe.
}
\examples{
## currently implemented metadata
metadata <- list(nirs_device = NULL,
                 nirs_columns = NULL,
                 sample_column = NULL,
                 event_column = NULL,
                 sample_rate = NULL,
                 event_sample_list = NULL,
                 fit_window = NULL,
                 display_window = NULL,
                 end_kinetics_window = NULL)

df <- data.frame(A = 1:3,
                 B = seq(10, 30, 10),
                 C = seq(11, 33, 11))
attributes(df)

nirs_data <- create_mnirs_data(
    df,
    metadata = list(nirs_columns = c("B", "C"),
                    sample_column = "A",
                    sample_rate = 1)
)
attributes(nirs_data)

}
