% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Apply filter methods to mNIRS data}
\usage{
filter_data(
  .data,
  sample_rate,
  handle_missing_values = c("interpolate", "omit", "none"),
  filter_method = "none",
  filter_parameters = NULL,
  ...
)
}
\arguments{
\item{.data}{A dataframe with mNIRS metadata.}

\item{sample_rate}{An integer scalar indicating the sample rate at which the
raw data was recorded (in Hz; samples per second).}

\item{handle_missing_values}{Indicates how to handle missing data globally
before filtering can be applied.
\itemize{
\item \emph{"interpolate"} (the default) will interpolate between existing values.
\item \emph{"omit"} will remove \code{NA}s before analysis and re-insert
them back to their original indices.
\item \emph{"none"} will pass through missing data. May cause errors.
}}

\item{filter_method}{A list of character vectors indicating the digital
filter method used to smooth each mNIRS data column, either column-wise
if the list is named by mNIRS columns, or globally if not named. See details.
\itemize{
\item \emph{"none"} (the default) will pass through raw data without any digital
filtering.
\item \emph{"smooth-spline"} fits a cubic smoothing spline to the supplied data with
and automatically determined smoothing parameter \code{spar}.
\item \emph{"low-pass"} uses a symmetrical (centred) butterworth low-pass filter.
\item \emph{"moving-average"} uses a symmetrical (centred) moving average filter.
}}

\item{filter_parameters}{(optional) Either \code{NULL} or a list of named
numeric vectors of parameters to pass through to the digital filter(s)
defined in \code{filter_method}. Either column-wise if the list is named by
mNIRS columns, or globally if not named. See details.}

\item{...}{Additional arguments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Applies digital filtering with either:
\enumerate{
\item A smoothing spline with \code{\link[stats:smooth.spline]{stats::smooth.spline()}}
\item A Butterworth low-pass filter with \code{\link[signal:butter]{signal::butter()}} and
\code{\link[=filtfilt_edges]{filtfilt_edges()}}
\item A simple moving average with \code{\link[zoo:rollapply]{zoo::rollapply()}}.
}
}
\details{
\code{handle_missing_values} must be defined before filtering can be performed.
\itemize{
\item \emph{"interpolate"} uses \code{\link[zoo:na.approx]{zoo::na.approx()}} to linearly interpolate across
missing data. Missing values on the edges of the dataframe will be
extrapolated from the first or last present value (see \emph{"rule = 2"} in
\code{\link[stats:approxfun]{stats::approx()}}). Additional arguments include \emph{"maxgap"} to
define a limit of consecutive \code{NA}s to fill. Longer gaps will be left
unchanged and may cause subsequent errors (see \emph{"maxgap"} in
\code{\link[zoo:na.approx]{zoo::na.approx()}}).
\item \emph{"omit"} will record the indices of \code{NA}s and omit them before
applying the digital filter. Then \code{NA}s are re-inserted back to their
original indices when the filtered dataframe is returned. CURRENTLY NOT
IMPLEMENTED.
\item \emph{"none"} will cause \emph{"low-pass"} or \emph{"smooth-spline"}
filtering to fail if there are nay missing or non-finite data.
}

\code{filter_method} may be a named list for column-wise filtering of individual
mNIRS columns. Or a non-named list or vector with one filtering method to
apply to all mNIRS columns. Currently will only work if either none or all
mNIRS columns are named. Such as:
\itemize{
\item \code{filter_method = c("smooth-spline")}.
\item \code{filter_method = list(SmO2 = "moving-average", HHb = "smooth-spline")}.
}

\code{filter_parameters} will be chosen by default to correspond to
\code{filter_method}. If \code{filter_method} is a named list, then
\code{filter_parameters} must be either unnamed for global parameters, or must
have all of the same mNIRS names as \code{filter_method}.
\itemize{
\item For \code{filter_method = "smooth-spline"} the default spar value is often good
enough. Higher values within the range \code{0} to \code{1} will apply more smoothing.
Lower values will apply less smoothing. CURRENTLY DEFAULT ONLY IMPLEMENTED.
\item For \code{filter_method = "low-pass"}, \code{filter_parameters = list(c(n, fc))},
where \code{n} is an integer scalar specifying the filter order number and
\code{fc} is a numeric scalar specifying the cutoff frequency in Hz.
\item For \code{filter_method = "moving-average"}, \code{filter_parameters = list(c(k))},
where \code{k} is an odd integer scalar specifying the window width (number of
samples) centred around the index sample (\code{x +- k/2-1}).
}
}
