% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.mNIRS.kinetics.R
\name{mNIRS.kinetics-methods}
\alias{mNIRS.kinetics-methods}
\alias{print.mNIRS.kinetics}
\alias{plot.mNIRS.kinetics}
\title{Methods for mNIRS.kinetics objects}
\usage{
\method{print}{mNIRS.kinetics}(x, ...)

\method{plot}{mNIRS.kinetics}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mNIRS.kinetics} returned from \code{\link[=process_kinetics]{process_kinetics()}}.}

\item{...}{Additional arguments (see \emph{Value}).}
}
\value{
Return:
\item{\code{print}}{Returns a model summary}
\item{\code{plot}}{Returns a \link[ggplot2:ggplot]{ggplot2} object}
}
\description{
Methods defined for objects returned from \code{\link[=process_kinetics]{process_kinetics()}}.
}
