% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.mNIRS.kinetics.R
\name{mNIRS.kinetics-methods}
\alias{mNIRS.kinetics-methods}
\alias{print.mNIRS.kinetics}
\alias{plot.mNIRS.kinetics}
\title{Methods for mNIRS.kinetics objects}
\usage{
\method{print}{mNIRS.kinetics}(x, ...)

\method{plot}{mNIRS.kinetics}(x, ...)
}
\arguments{
\item{x}{Object of class \code{mNIRS.kinetics} returned from \code{\link[=process_kinetics]{process_kinetics()}}.}

\item{...}{Additional arguments.
\describe{
\item{\code{plot_coefs}}{\code{plot_coefs = TRUE} will display kinetics coefficients
at the top left of the plot.}
\item{\code{plot_diagnostics}}{\code{plot_diagnostics = TRUE} will display kinetics
model diagnostic metrics at the bottom right of the plot.}
\item{\code{plot_residuals}}{\code{plot_diagnostics = TRUE} will display kinetics
model residuals along the bottom of the plot. Residual values are correctly
scaled about the zero dotted line}
}}
}
\value{
Return:
\item{\code{print}}{Returns a model summary}
\item{\code{plot}}{Returns a \link[ggplot2:ggplot]{ggplot2} object}
}
\description{
Generic methods defined for objects returned from \code{\link[=process_kinetics]{process_kinetics()}}.
}
