% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mNIRS.data.R
\name{mNIRS_palette}
\alias{mNIRS_palette}
\title{Custom mNIRS colour palette}
\usage{
mNIRS_palette(n = NULL)
}
\arguments{
\item{n}{Number of colours to return, in order.}
}
\value{
Named character vector of hex colours.
}
\description{
Custom mNIRS colour palette
}
\examples{
\dontrun{
scales::show_col(mNIRS_palette())
}

}
