% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoexponential.R
\name{monoexponential}
\alias{monoexponential}
\title{Monoexponential Function}
\usage{
monoexponential(x, A, B, TD, tau)
}
\arguments{
\item{x}{A numeric vector for the predictor variable at which to calculate
the response variable.}

\item{A}{A numeric parameter for the starting value.}

\item{B}{A numeric parameter for the ending asymptote value.}

\item{TD}{A numeric parameter for the time delay in the units of \code{x}
before inflection.}

\item{tau}{A numeric parameter for the time constant of the exponential
inflection in the units of \code{x}.}
}
\value{
A numeric vector of fitted values from a monoexponential function.
}
\description{
A function to calculate a four-parameter monoexponential curve.
}
\examples{
set.seed(13)
x <- seq(0, 60, by = 2)
A <- 10
B <- 100
TD <- 15
tau <- 8
y <- monoexponential(x, A, B, TD, tau) + rnorm(length(x), 0, 3)
data <- data.frame(x, y)

model <- nls(y ~ SSmonoexp(x, A, B, TD, tau), data = data)
model
fitted <- fitted(model)

plot(x, y)
lines(x, y)
points(x, y)
lines(x, fitted, col = "red")

}
