% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_data.R
\name{normalise_data}
\alias{normalise_data}
\title{Normalise Signal Range}
\usage{
normalise_data(
  x,
  shift_to = 0,
  normalise_range = c(0, 100),
  shift_by,
  shift_position = c("minimum", "start value", "maximum"),
  mean_samples,
  ...
)
}
\arguments{
\item{...}{Additional arguments.}
}
\value{
A numeric vector of filtered data.
}
\description{
Shift the signal values while preserving the dynamic range (e.g. shift to
positive values), or normalise the signal to within a new dynamic range (e.g.
normalise range to \verb{[0, 100]}).
}
\details{
...
}
