% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mNIRS.data.R
\name{plot.mNIRS.data}
\alias{plot.mNIRS.data}
\title{Plot mNIRS.data objects}
\usage{
\method{plot}{mNIRS.data}(data, ...)
}
\arguments{
\item{data}{object of class \code{mNIRS.data} as returned from \code{\link[=create_mNIRS_data]{create_mNIRS_data()}}}

\item{...}{Additional arguments (\emph{currently not used}).}
}
\value{
A \link[ggplot2:ggplot]{ggplot2} object.
}
\description{
Create a simple plot for objects returned from \code{\link[=create_mNIRS_data]{create_mNIRS_data()}}.
}
