% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.mNIRS.data.R
\name{plot.mNIRS.data}
\alias{plot.mNIRS.data}
\title{Methods for mNIRS.data objects}
\usage{
\method{plot}{mNIRS.data}(x, ...)
}
\arguments{
\item{x}{object of class \code{mNIRS.data} as returned from \code{\link[=create_mnirs_data]{create_mnirs_data()}}}

\item{...}{further arguments passed through, see description of return value
for details.}
}
\value{
\describe{
\item{\code{plot}}{Returns a simple plot of mNIRS data with
\link[ggplot2:ggplot]{ggplot2}.}
}
}
\description{
Methods defined for objects returned from \code{\link[=create_mnirs_data]{create_mnirs_data()}}.
}
