% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data.R
\name{prepare_kinetics_data}
\alias{prepare_kinetics_data}
\title{Prepare mNIRS Data for Kinetics Analysis}
\usage{
prepare_kinetics_data(
  data,
  event_index = NULL,
  event_sample = NULL,
  event_label = NULL,
  fit_window = c(30, 180),
  display_window = NULL,
  group_events = list("distinct", "ensemble"),
  ...
)
}
\arguments{
\item{data}{A dataframe.}

\item{event_index}{(\emph{Optional}). A numeric vector indicating the start
indices of kinetics events. Any of \code{event_label}, \code{event_index}, or
\code{event_sample} will be passed along for further analysis.}

\item{event_sample}{(\emph{Optional}). A vector in the same class as
\code{sample_column} indicating the start of kinetic events.}

\item{event_label}{(\emph{Optional}). A character vector specified in
\code{event_column} indicating the start of kinetics events.}

\item{fit_window}{A two-element numeric vector defining the number of
samples around the kinetics events to include in the fitting process (see
\emph{Details}).}

\item{display_window}{(\emph{Optional}). A two-element numeric vector defining
the number of samples around the kinetics events to include for display (see
\emph{Details}).}

\item{group_events}{Indicates how kinetics events should be analysed
(see \emph{Details}).}

\item{...}{Additional arguments.}
}
\value{
A list of \link[tibble:tibble-package]{tibbles} of class \code{mNIRS.data}
with metadata available with \code{attributes()}.
}
\description{
Processes a dataframe for kinetics analysis.
}
\details{
\code{fit_window} is a two-element vector \code{c(before, after)} specifying the number of
samples before and after the kinetics event to include in the kinetics fitting
process.

\code{display_window} is a two-element vector \code{c(before, after)} specifying the
number of samples before and after the kinetics event to include in the prepared
kinetics dataframes, to allow for plotting of data beyond the bounds of the
fitting process.

\code{group_events} indicates how kinetics events should be analysed, either
separately, or grouped and ensemble averaged similar to oxygen uptake kinetics.

\describe{
\item{\code{group_events = "distinct"}}{Will prepare a list of unique dataframes
for each kinetics event (\emph{default}).}
\item{\code{group_events = "ensemble"}}{Will prepare one dataframe with the
ensemble-averaged data from all mNIRS kinetics events.}
\item{\code{group_events = list(c(1, 2), c(3, 4))}}{Will group kinetic events
together in sequence of appearance, and prepare a list of ensemble-averaged
dataframes for each group. Any kinetic events detected in the data but
not explicitly defined here will return as un-grouped dataframes.}
}
}
