% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data.R
\name{prepare_kinetics_data}
\alias{prepare_kinetics_data}
\title{Prepare Dataframe for mNIRS Kinetics Analysis}
\usage{
prepare_kinetics_data(
  .data,
  multiple_kinetics_events = c("distinct", "ensemble"),
  baseline_fit_window = 30,
  kinetics_fit_window = 240,
  baseline_display_window = 30,
  kinetics_display_window = 240,
  end_kinetics_window = NULL,
  ...
)
}
\arguments{
\item{.data}{The data retrived \code{\link[=process_data]{process_data()}}.}

\item{multiple_kinetics_events}{Indicates whether each \emph{"distinct"} kinetics
events should be analysed separately (the default). Or whether multiple
kinetics events should be \emph{"ensemble"} averaged and analysed together
(similar to pulmonary VO2 kinetics).}

\item{baseline_fit_window}{A numeric scalar specifying the number of
samples preceding to the kinetics events to include as baseline.}

\item{kinetics_fit_window}{A numeric scalar specifying the number of
samples following to the kinetics events to include for kinetics modelling.}

\item{baseline_display_window}{A numeric scalar specifying the number of
samples preceding to the kinetics events to include for display (will not
necessarily be modelled).}

\item{kinetics_display_window}{A numeric scalar specifying the number of
samples following to the kinetics events to include for display (will not
necessarily be modelled).}

\item{end_kinetics_window}{A numeric scalar specifying the number of
samples in which to look for a peak or nadir value indicating the kinetics
plateau.}

\item{...}{Additional arguments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
This function retrieves processed data from \code{\link[=process_data]{process_data()}} with manually
identified kinetics events. Kinetics parameters are defined to extract
data excerpts to fit those kinetics events. Each kinetics event can be
analysed separately, or ensemble-averaged and analysed together.
}
\details{
TODO
}
