% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data.R
\name{prepare_kinetics_data}
\alias{prepare_kinetics_data}
\title{Prepare mNIRS Data for Kinetics Analysis}
\usage{
prepare_kinetics_data(
  data,
  event_sample = NULL,
  event_label = NULL,
  event_index = NULL,
  fit_window = c(30, 180),
  display_window = NULL,
  peak_window = 30,
  group_events = list("distinct", "ensemble"),
  nirs_columns = NULL,
  sample_column = NULL,
  event_column = NULL,
  sample_rate = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe of class \code{"mNIRS.data"}.}

\item{event_sample}{An \emph{optional} numeric vector corresponding to values of
\code{sample_column} indicating the start of kinetic events. i.e., by time value
or sample number.}

\item{event_label}{An \emph{optional} character vector corresponding to values of
\code{event_column} indicating the start of kinetics events. i.e., by an event
label such as \emph{"end work"}.}

\item{event_index}{An \emph{optional} numeric vector indicating the starting row
indices of kinetics events. i.e., to identify the start of kinetic events by
row number.}

\item{fit_window}{A two-element numeric vector in the form \code{c(before, after)}
in units of \code{sample_column}, defining the window around the kinetics events
to include in the model fitting process (\emph{default = c(30, 180)}).}

\item{display_window}{(\emph{Not currently implemented})
An \emph{optional} two-element numeric vector in the form
\code{c(before, after)} in units of \code{sample_column}, defining the window around
the kinetics events to include for display, but not for model fitting.}

\item{peak_window}{A numeric scalar indicating the local window in units of
\code{sample_column} after the kinetics peak (trough) value to look for subsequent
higher (lower) values. The kinetics model will be fit to the data up to the
first local peak (trough) value with no subsequent higher (lower) values
within the lesser of either the \code{peak_window} or the limits of \code{fit_window}.}

\item{group_events}{Indicates how kinetics events should be analysed. Typically
either \emph{"distinct"} (\emph{the default}) or \emph{"ensemble"}, but can be manually
specified (see \emph{Details}).}

\item{nirs_columns}{A character vector indicating the mNIRS data columns to
be processed from your dataframe. Must match \code{data} column names exactly.
Will be taken from metadata if not defined explicitly.}

\item{sample_column}{An \emph{optional} character scalar indicating the name of
a time or sample data column. Must match exactly. Defining
column names explicitly will overwrite metadata.}

\item{event_column}{An \emph{optional} character scalar indicating the name of
an event or lap data column. Must match exactly. Defining
column names explicitly will overwrite metadata.}

\item{sample_rate}{A numeric scalar for the sample rate in Hz. Will be taken
from metadata if not defined explicitly.}

\item{...}{Additional arguments.}
}
\value{
A list of \link[tibble:tibble-package]{tibbles} of class \code{mNIRS.data}
with metadata available with \code{attributes()}.
}
\description{
Processes a list of one or more dataframes of class \code{"mNIRS.data"} representing
distinct or ensembled kinetics events for further analysis.
}
\details{
\code{display_window} defines the widest range of data before and after the kinetics
event which will be passed on in the dataframe, but not necessarily included
in the modelling process. \code{fit_window} defines the widest extent of data
before and after the kinetics event which will be included in the modelling
process. \code{peak_window} limits the extent of data included to be modelled after
the kinetics event to the first local peak (trough) value of the response
variable (i.e. \code{y}), with no higher (lower) values within that window.

\code{group_events} indicates how kinetics events should be analysed, either
separately, or grouped and ensemble averaged similar to oxygen uptake kinetics.

\describe{
\item{\code{group_events = "distinct"}}{Will prepare a list of unique dataframes
for each kinetics event (\emph{default}).}
\item{\code{group_events = "ensemble"}}{Will prepare one dataframe with the
ensemble-averaged data from all mNIRS kinetics events.}
\item{\code{group_events = list(c(1, 2), c(3, 4))}}{Will group kinetic events
together in sequence of appearance, and prepare a list of ensemble-averaged
dataframes for each group. Any kinetic events detected in the data but
not explicitly defined here will return as a distinct dataframe.}
}
}
