% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data.R
\name{prepare_kinetics_data}
\alias{prepare_kinetics_data}
\title{Prepare mNIRS Data for Kinetics Analysis}
\usage{
prepare_kinetics_data(
  data,
  event_index = NULL,
  event_sample = NULL,
  event_label = NULL,
  fit_baseline_window = 30,
  fit_kinetics_window = 180,
  display_baseline_window = NULL,
  display_kinetics_window = NULL,
  group_kinetics_events = list("distinct", "ensemble"),
  ...
)
}
\arguments{
\item{data}{A dataframe containing mNIRS data.}

\item{event_index}{(\emph{optional}) A numeric vector indicating the start
indices of kinetics events.
\itemize{
\item Any of \code{event_label}, \code{event_index}, or \code{event_sample} will be passed
along for further analysis.
}}

\item{event_sample}{(\emph{optional}) A vector in the same class as
\code{sample_column} indicating the start of kinetic events.}

\item{event_label}{(\emph{optional}) A character vector specified in
\code{event_column} indicating the start of kinetics events.}

\item{fit_baseline_window}{A numeric scalar specifying the number of
samples preceding to the kinetics events to include as baseline.}

\item{fit_kinetics_window}{A numeric scalar specifying the number of
samples following to the kinetics events to include for kinetics modelling.}

\item{display_baseline_window}{A numeric scalar specifying the number of
samples preceding to the kinetics events to include for display.}

\item{display_kinetics_window}{A numeric scalar specifying the number of
samples following to the kinetics events to include for display.}

\item{group_kinetics_events}{Indicates whether kinetics events should
be analysed separately or together.
\itemize{
\item \code{"distinct"} (\emph{the default}) will prepare a list of unique dataframes for
each kinetics event.
\item `"ensemble"`` will prepare one dataframe with the ensemble-averaged mNIRS
data from all kinetics events (similar to pulmonary VO2 kinetics).
\item Alternatively, a list of numeric vectors can be specified to
ensemble-average groups of kinetics events together. Such as
\code{group_kinetics_events = list(c(1, 2), c(3, 4))} for two sets of
repeated bouts. Any bout numbers not listed explicitly will return
un-grouped dataframes.
}}

\item{...}{Additional arguments.}
}
\value{
A list of \link[tibble:tibble-package]{tibbles} of class \code{mNIRS.data}
with metadata available with \code{attributes()}.
}
\description{
Processes a dataframe for kinetics analysis
}
\details{
...
}
