% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_kinetics_data.R
\name{prepare_kinetics_data}
\alias{prepare_kinetics_data}
\title{Prepare mNIRS Data for Kinetics Analysis}
\usage{
prepare_kinetics_data(
  data,
  event_sample = NULL,
  event_label = NULL,
  event_index = NULL,
  fit_span = c(30, 180),
  display_span = NULL,
  group_events = list("distinct", "ensemble"),
  nirs_channels = NULL,
  sample_channel = NULL,
  event_channel = NULL,
  sample_rate = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe of class \code{"mNIRS.data"}.}

\item{event_sample}{An \emph{optional} numeric vector corresponding to values of
\code{sample_channel} indicating the start of kinetic events. i.e., by time value
or sample number.}

\item{event_label}{An \emph{optional} character vector corresponding to values of
\code{event_channel} indicating the start of kinetics events. i.e., by an event
label such as \emph{"end work"}.}

\item{event_index}{An \emph{optional} numeric vector indicating the starting row
indices of kinetics events. i.e., to identify the start of kinetic events by
row number.}

\item{fit_span}{A two-element numeric vector in the form \code{c(before, after)}
in units of \code{sample_channel}, defining the window around the kinetics events
to include in the model fitting process (\emph{default} \code{fit_span = c(30, 180)}).}

\item{display_span}{\emph{\verb{<under development>}}.}

\item{group_events}{Indicates how kinetics events should be analysed. Typically
either \emph{"distinct"} (the \emph{default}) or \emph{"ensemble"}, but can be manually
specified (see \emph{Details}).}

\item{nirs_channels}{A character vector indicating the mNIRS data channels to
be processed from your dataframe. Must match \code{data} column names exactly.
Will be taken from metadata if not defined explicitly.}

\item{sample_channel}{A character scalar indicating the time or sample data
channel. Must match \code{data} column names exactly. Will be taken from metadata
if not defined explicitly.}

\item{event_channel}{An \emph{optional} character scalar indicating an event or
lap data channel. Must match \code{data} column names exactly. Will be taken from
metadata if not defined explicitly.}

\item{sample_rate}{A numeric scalar for the sample rate in Hz. Will be taken
from metadata if not defined explicitly.}

\item{...}{Additional arguments.}
}
\value{
A list of \link[tibble:tibble-package]{tibbles} of class \code{"mNIRS.data"}
with metadata available with \code{attributes()}.
}
\description{
Processes a list of one or more dataframes of class \code{"mNIRS.data"} representing
distinct or ensembled kinetics events for further analysis.
}
\details{
\code{display_span} defines the widest range of data before and after the kinetics
event which will be passed on in the dataframe, but not included in the
modelling process. \code{fit_span} defines the widest extent of data before and
after the kinetics event which may be included in the modelling process.

\code{group_events} indicates how kinetics events should be analysed, either
separately, or grouped and ensemble averaged similar to oxygen uptake kinetics.
\describe{
\item{\verb{group_events = }\emph{\code{"distinct"}}}{Will prepare a list of unique dataframes
for each kinetics event (\emph{default}).}
\item{\verb{group_events = }\emph{\code{"ensemble"}}}{Will prepare one dataframe with the
ensemble-averaged data from all mNIRS kinetics events.}
\item{\code{group_events = list(c(1, 2), c(3, 4))}}{Will group kinetic events
together in sequence of appearance, and prepare a list of ensemble-averaged
dataframes for each group. Any kinetic events detected in the data but
not explicitly defined here will return as a distinct dataframe.}
}
}
