% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_data}
\alias{read_data}
\title{Read mNIRS Data From File}
\usage{
read_data(
  file_path,
  nirs_columns,
  sample_column = NULL,
  event_column = NULL,
  sample_rate = NULL,
  .numeric_time = TRUE,
  .keep_all = FALSE,
  .verbose = TRUE
)
}
\arguments{
\item{file_path}{The file path including extension (either \emph{".xlsx"},
\emph{".xls"}, or \emph{".csv"}) to import.}

\item{nirs_columns}{A character vector indicating the mNIRS data columns
to import from the target file. Must match exactly. A named character vector
can be used to rename columns (see \emph{Details}).}

\item{sample_column}{(\emph{Optional}). A character scalar indicating the name of
a time or sample data column. Must match exactly. A named character vector
can be used to rename columns.}

\item{event_column}{(\emph{Optional}). A character scalar indicating the name of
an event or lap data column. Must match exactly. A named character vector
can be used to rename columns.}

\item{sample_rate}{(\emph{Optional}). A numeric scalar for the sample rate in Hz.
If not defined explicitly, will be estimated from the file data
(see \emph{Details}).}

\item{.numeric_time}{A logical. \code{TRUE} (\emph{default}) will convert
date-time formatted columns to numeric values in seconds. \code{FALSE} will retain
these columns formatted as date-time in the format of the original file.}

\item{.keep_all}{A logical. \code{FALSE} (\emph{default}) will only include the
explicitly indicated data columns. \code{TRUE} will include all columns detected
from the file.}

\item{.verbose}{A logical. \code{TRUE} (\emph{default}) will return warnings and
messages which can be used for data error checking. \code{FALSE} will silence these
messages. Errors will always be returned.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Reads mNIRS data from file, and returns a dataframe.
}
\details{
Column names are matched to a single row, anywhere in the data file, not
necessarily the top row of the file. Column names must match exactly to data
file text. If there are duplicate column names, the columns will be matched in
order of appearance and appended with \code{...X}, where \code{X} is the column number.
You may want to double check the correct columns have been assigned as intended.

Columns can be renamed in the format \code{c("new_name" = "original_name")}, where
\code{"original_name"} should match the file text exactly.

\code{sample_rate} is required for certain \code{{mNIRS}} functions. If it is not
defined explicitly, \code{sample_rate} will be estimated based
on the mean difference between values in the \code{sample_column}. If
\code{sample_column} is not defined, then \code{sample_rate} will be set to 1 Hz.
If \code{sample_column} in the data file contains integer row numbers, then
\code{sample_rate} will be incorrectly estimated to be 1 Hz, and should be
defined explicitly.

Columns and rows which are entirely missing (\code{NA}) are omitted.
}
