% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dataframe.R
\name{resample_dataframe}
\alias{resample_dataframe}
\title{Resample Data}
\usage{
resample_dataframe(
  data,
  sample_column = NULL,
  sample_rate = NULL,
  resample_rate = NULL,
  resample_time = NULL,
  ...
)
}
\arguments{
\item{data}{A dataframe containing mNIRS data.}

\item{sample_column}{A character scalar indicating the name of
the time or sample data column. Must match exactly.}

\item{sample_rate}{An integer scalar for the sample rate in Hz of the
dataframe.}

\item{resample_rate}{An integer scalar indicating the desired output sample
rate (in Hz) to convert the file to.}

\item{resample_time}{A numeric scalar indicating the desired sample time
(in seconds) to convert the file to.}

\item{...}{Additional arguments (currently not used).}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Resamples mNIRS data using time-weighted interpolation.
}
\details{
\code{sample_column} will be taken from metadata for an mNIRS dataframe.

If not defined explicitly, \code{sample_rate} will be estimated based
on the mean difference between values in the \code{sample_column}. If
\code{sample_column} contains integer sample numbers rather than time values,
then \code{sample_rate} will be incorrectly estimated to be 1 Hz, and should
be defined explicitly.
}
