% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_dataframe_TODO.R
\name{resample_dataframe}
\alias{resample_dataframe}
\title{Resample Data}
\usage{
resample_dataframe(data, sample_rate = 1, resample_rate = 1, resample_by)
}
\arguments{
\item{data}{A dataframe containing mNIRS data.}

\item{...}{Additional arguments.}
}
\value{
A \link[tibble:tibble-package]{tibble}.
}
\description{
Down-samples or up-samples mNIRS dataframe
}
\details{
Down-sampling uses
}
