% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rescale_data.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescale Data Range}
\usage{
rescale_data(data, nirs_columns = list(), rescale_range = c(0, 100))
}
\arguments{
\item{data}{A dataframe.}

\item{nirs_columns}{A \code{list()} of character vectors indicating the column
names for data channels to be rescaled (see \emph{Details}).}

\item{rescale_range}{A numeric vector in the form \code{c(min, max)},
indicating the range of output values to which data channels will be rescaled.}
}
\value{
A \link[tibble:tibble-package]{tibble} of class \code{mNIRS.data} with
metadata available with \code{attributes()}.
}
\description{
Rescale the range of data channels to a new dynamic range, while preserving
the relative scaling across columns. e.g. rescale the range of data to
\code{c(0,100)}.
}
\details{
\code{nirs_columns = list()} can be used to group data columns to preserve
relative scaling. channels grouped together in a vector will preserve
relative scaling across channels. Should match column names in the dataframe
exactly.
\describe{
\item{\code{nirs_columns = list("A", "B", "C")}}{will rescale each column
separately. Relative scaling will be lost between data channels.}
\item{\code{nirs_columns = list(c("A", "B", "C"))}}{will rescale all columns
together. Relative scaling is preserved across the group of data channels.}
\item{\code{nirs_columns = list(c("A", "B"), c("C", "D"))}}{will rescale columns
\code{A} and \code{B} together, and columns \code{C} and \code{D} together. Relative scaling
is preserved within each group, but not across groups of data channels.}
}
}
