% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_slope.R
\name{rolling_slope}
\alias{rolling_slope}
\title{Calculate Rolling Slope}
\usage{
rolling_slope(
  y,
  x = seq_along(y),
  width,
  align = c("center", "left", "right"),
  na.rm = FALSE
)
}
\arguments{
\item{y}{A numeric vector of the response variable.}

\item{x}{A numeric vector of the predictor variable, defaults to using the
index of \code{x = seq_along(y)}.}

\item{width}{A numeric scalar defining the window width in units of \code{x}
for rolling calculations.}

\item{align}{Specifies the window alignment of \code{width} as \emph{"center"}
(the \emph{default}), \emph{"left"}, or \emph{"right"}. Where \emph{"left"} is \emph{forward looking},
and \emph{"right"} is \emph{backward looking} by the window \code{width} from the current
sample.}

\item{na.rm}{A logical indicating how missing data will be handled. \code{FALSE}
(the \emph{default}) will perform complete case analysis and return the rolling
slopes where all local \code{y} samples are valid. \code{TRUE} will return the rolling
slopes where the local target \code{y} sample is valid.}
}
\value{
A numeric vector of rolling local slopes in units of \code{y/x} of the
same length as \code{y}.
}
\description{
Computes rolling first derivative (linear regression slope) for a numeric
response variable \code{y} within a moving window.
}
\details{
Uses the least squares formula on complete case analysis to calculate
local slopes within a rolling window along \code{x} specified by \code{width} in units
of \code{x}. i.e. \code{x = 10} would include samples within a 10-second window for
time series data.
}
\examples{
y <- c(1, 3, 2, 5, 8, 7, 9, 12, 11, 15, 14, 17, 18)
rolling_slope(y, width = 3)
rolling_slope(y, width = 3, align = "left")

y_na <- c(1, 3, NA, 5, 8, 7, 9, 12, NA, NA, NA, 17, 18)
rolling_slope(y_na, width = 3)

}
\seealso{
\code{\link[zoo:rollapply]{zoo::rollapply()}}
}
