% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mNIRS.data.R
\name{scale_colour_mNIRS}
\alias{scale_colour_mNIRS}
\alias{scale_fill_mNIRS}
\title{Scales for custom mNIRS palette}
\usage{
scale_colour_mNIRS(...)

scale_fill_mNIRS(...)
}
\arguments{
\item{...}{Arguments passed to discrete_scale.}
}
\description{
Scales for custom mNIRS palette
}
\seealso{
\code{\link[=mNIRS_palette]{mNIRS_palette()}}
}
