% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_slope.R
\name{slope}
\alias{slope}
\title{Determine the Slope of a Numeric Vector}
\usage{
slope(y, x = NULL, na.rm = FALSE)
}
\arguments{
\item{y}{A numeric vector of the response variable.}

\item{x}{An \emph{optional} numeric vector of the predictor variable.
If \code{x = NULL}, uses \code{x = seq_along(y)}.}

\item{na.rm}{A logical indicating whether to exclude NA values from rolling
slope calculations.}
}
\value{
A numeric slope value.
}
\description{
Computes the rate of change for a numeric response variable \code{y}.
}
\keyword{internal}
