% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monoexponential.R
\name{SSmonoexp}
\alias{SSmonoexp}
\title{Self-Starting \code{nls} Monoexponential Model}
\usage{
SSmonoexp(x, A, B, TD, tau)
}
\arguments{
\item{x}{A numeric vector at which to evaluate the model.}

\item{A}{A numeric parameter for the starting value.}

\item{B}{A numeric parameter for the ending asymptote value.}

\item{TD}{A numeric parameter for the time delay in the units of \code{x}
before inflection.}

\item{tau}{A numeric parameter for the time constant of the exponential
inflection in the units of \code{x}.}
}
\value{
A numeric vector of the same length as \code{x}. The value of the
expression \code{ifelse(x <= TD, A, A + (B - A) * (1 - exp((TD - x) / tau)))}.
}
\description{
This \code{selfStart} model evaluates the four-parameter
\link[=monoexponential]{monoexponential} function. It creates initial
estimates for the parameters \code{A}, \code{B}, \code{TD}, and \code{tau}.
}
\examples{
set.seed(13)
x <- seq(0, 60, by = 2)
A <- 10
B <- 100
TD <- 15
tau <- 8
y <- monoexponential(x, A, B, TD, tau) + rnorm(length(x), 0, 3)
data <- data.frame(x, y)

model <- nls(y ~ SSmonoexp(x, A, B, TD, tau), data = data)
model
fitted <- fitted(model)

plot(x, y)
lines(x, y)
points(x, y)
lines(x, fitted, col = "red")

}
\seealso{
\link[stats:nls]{nls}, \link[stats:selfStart]{selfStart}
}
