% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_mNIRS.R
\name{theme_mNIRS}
\alias{theme_mNIRS}
\title{mNIRS ggplot2 theme}
\usage{
theme_mNIRS(
  base_size = 14,
  base_family = "sans",
  border = c("partial", "full"),
  ...
)
}
\arguments{
\item{base_size}{Base font size, given in pts.}

\item{base_family}{Base font family.}

\item{border}{Define either a \emph{partial} or \emph{full} border around plots.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2} object.
}
\description{
A \link[ggplot2:ggplot2-package]{ggplot2} theme for pretty display.
}
