% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_fixed_coefs.R
\name{update_fixed_coefs}
\alias{update_fixed_coefs}
\title{Update a model object with Fixed coefficients}
\usage{
update_fixed_coefs(model, ...)
}
\arguments{
\item{model}{An existing model object from \code{lm}, \code{nls}, \code{glm}, and many others.}

\item{...}{Additional arguments to define a fixed coefficients.}
}
\value{
A model object.
}
\description{
Re-fit a model with fixed coefficients provided as an argument. Fixed
coefficients are not modified when optimising for best fit.
}
\details{
If additional arguments are left blank, or if a coefficient does not exist in
the model, the model will be returned unchanged.
}
\keyword{internal}
